[Refactor] : database model 분리
Git : #23
