+++ eb_fast_api/database/sources/model/models.py
@@ -1,0 +1,3 @@
+from eb_fast_api.database.sources.model.user import User
+from eb_fast_api.database.sources.model.schedule import Schedule
+from eb_fast_api.database.sources.model.place import Place
