+++ eb_fast_api/domain/schema/sources/schema.py
@@ -1,9 +1,8 @@
 from pydantic import BaseModel
-from eb_fast_api.database.sources.model import User
 from eb_fast_api.snippets.sources import pwdcrypt
 from typing import Optional
 from datetime import datetime
-from eb_fast_api.database.sources.model import Schedule, Place, User
+from eb_fast_api.database.sources.model.models import User, Schedule, Place
 
 
 class Coordi(BaseModel):
